+++ commands/dicdel.js
@@ -22,7 +22,7 @@ module.exports = {
 
     const connection = app.connections_map.get(guild_id);
 
-    const server_file = app.bot_utils.get_server_file(guild_id);
+    const server_file = app.data_utils.get_server_file(guild_id);
     let dict = server_file.dict;
 
     const target = interaction.options.get("target").value;
@@ -43,7 +43,7 @@ module.exports = {
 
     dict = dict.filter(word => word[0] !== target);
 
-    app.bot_utils.write_serverinfo(guild_id, server_file, { dict: dict });
+    app.data_utils.write_serverinfo(guild_id, server_file, { dict: dict });
 
     if(connection) connection.dict = dict;
