+++ commands/removeautojoin.js
@@ -27,7 +27,7 @@ module.exports = {
 
     const guild_id = interaction.guild.id;
 
-    const autojoin_list = app.bot_utils.get_autojoin_list();
+    const autojoin_list = app.data_utils.get_autojoin_list();
 
     if(!autojoin_list[guild_id] || !autojoin_list[guild_id][voice_channel_id]){
       await interaction.reply({ content:'設定がないよ', ephemeral: true });
@@ -36,7 +36,7 @@ module.exports = {
 
     delete autojoin_list[guild_id][voice_channel_id];
 
-    app.bot_utils.write_autojoin_list(autojoin_list);
+    app.data_utils.write_autojoin_list(autojoin_list);
     app.setup_autojoin();
 
     await interaction.reply({ content: `自動接続を設定しました！`, ephemeral: true });
