+++ src/bot_utils.js
@@ -1,10 +1,9 @@
-const fs = require('fs');
 const crypto = require('crypto');
 
 const ResurrectionSpell = require('./resurrection_spell.js');
 const SafeRegexpUtils = require('./safe_regexp_utils.js');
 
-const { SERVER_DIR, EXTEND_PASS, IS_PONKOTSU } = require('../config.json');
+const { EXTEND_PASS } = require('../config.json');
 
 const { shortcut } = require('../shortcuts.json');
 
@@ -21,17 +20,6 @@ module.exports = class BotUtils{
   #EXTEND_ENABLE;
   #VOICE_REGEXP_NAME;
   #voice_list;
-  #autojoin_cache;
-  #uservoices_cache;
-
-  #DEFAULT_SETTING = {
-    user_voices: {
-      DEFAULT: { voice: 1, speed: 100, pitch: 100, intonation: 100, volume: 100, is_force_server: false }
-    },
-    dict: [["Discord", "でぃすこーど", 2], ["さんが退出しました", "さんが射出されました", 2]],
-    is_ponkotsu: !!IS_PONKOTSU
-  }
-  #SETTING_LISTS;
 
   constructor(logger){
     this.#logger = logger;
@@ -39,17 +27,9 @@ module.exports = class BotUtils{
     this.#VOICE_REGEXP_SPELL = new RegExp(`[${ResurrectionSpell.spell_chars()}]+`, 'g');
 
     this.#EXTEND_ENABLE = EXTEND_PASS !== undefined && EXTEND_PASS !== "none";
-
-    this.#autojoin_cache = {};
-    this.#uservoices_cache = {};
   }
 
   init_voicelist(voice_list, voice_liblary_list){
-    // デフォルトのボイスIDはエンジンラッパー提供のボイスリストの0番目
-    // つまりエンジン0のボイス0ってこと
-    this.#DEFAULT_SETTING.user_voices.DEFAULT.voice = voice_list[0].value;
-    this.#SETTING_LISTS = Object.keys(this.#DEFAULT_SETTING);
-
     const list = voice_list.toSorted((a, b) => a.value - b.value);
 
     let add = [];
@@ -152,100 +132,5 @@ module.exports = class BotUtils{
     this.#logger.debug(`Pass = ${pass}, Command = ${extend_command[1]}`);
 
     return extend_command[1] === pass;
-  }
-
-  get_server_file(guild_id){
-    let result = this.#DEFAULT_SETTING;
-
-    if(fs.existsSync(`${SERVER_DIR}/${guild_id}.json`)){
-      try{
-        let json = JSON.parse(fs.readFileSync(`${SERVER_DIR}/${guild_id}.json`));
-
-        for(let l of this.#SETTING_LISTS){
-          if(json[l] === undefined){
-            json[l] = this.#DEFAULT_SETTING[l];
-            continue;
-          }
-        }
-
-        result = json;
-
-        this.#logger.debug(`loaded server conf: ${JSON.stringify(result, null, "  ")}`);
-      }catch(e){
-        this.#logger.info(e);
-        result = this.#DEFAULT_SETTING;
-      }
-    }
-
-    return JSON.parse(JSON.stringify(result));
-  }
-
-  write_serverinfo(guild_id, from, update){
-    let result = {};
-    for(let l of this.#SETTING_LISTS){
-      if(l in update) result[l] = update[l];
-      else if(l in from) result[l] = from[l];
-      else result[l] = this.#DEFAULT_SETTING[l];
-    }
-
-    try{
-      fs.writeFileSync(`${SERVER_DIR}/${guild_id}.json`, JSON.stringify(result, null, "  "));
-    }catch(e){
-      this.#logger.info(e);
-    }
-  }
-
-  get_autojoin_list(){
-    if(Object.keys(this.#autojoin_cache).length){
-      return JSON.parse(JSON.stringify(this.#autojoin_cache));
-    }
-
-    let result = {};
-    try{
-      let json = JSON.parse(fs.readFileSync(`${SERVER_DIR}/autojoin.json`));
-
-      result = json;
-    }catch(e){
-      this.#logger.info(e);
-      result = {};
-    }
-
-    return result;
-  }
-
-  write_autojoin_list(list){
-    try{
-      fs.writeFileSync(`${SERVER_DIR}/autojoin.json`, JSON.stringify(list, null, "  "));
-      this.#autojoin_cache = JSON.parse(JSON.stringify(list));
-    }catch(e){
-      this.#logger.info(e);
-    }
-  }
-
-  get_uservoices_list(){
-    if(Object.keys(this.#uservoices_cache).length){
-      return JSON.parse(JSON.stringify(this.#uservoices_cache));
-    }
-
-    let result = {};
-    try{
-      let json = JSON.parse(fs.readFileSync(`${SERVER_DIR}/uservoices.json`));
-
-      result = json;
-    }catch(e){
-      this.#logger.info(e);
-      result = {};
-    }
-
-    return result;
-  }
-
-  write_uservoices_list(list){
-    try{
-      fs.writeFileSync(`${SERVER_DIR}/uservoices.json`, JSON.stringify(list, null, "  "));
-      this.#uservoices_cache = JSON.parse(JSON.stringify(list));
-    }catch(e){
-      this.#logger.info(e);
-    }
   }
 }
