+++ src/setvoice.js
@@ -9,7 +9,7 @@ module.exports = async (interaction, type) => {
   const global_voice = app.uservoices_map.get(member_id);
   if(!!global_voice && global_voice.enabled) is_global_uservoice = true;
 
-  const server_file = app.bot_utils.get_server_file(guild_id);
+  const server_file = app.data_utils.get_server_file(guild_id);
 
   if(!!server_file.user_voices[member_id]?.is_force_server) is_global_uservoice = false;
 
@@ -17,7 +17,7 @@ module.exports = async (interaction, type) => {
   if(!is_global_uservoice || type === 'is_force_server'){
     voices = server_file.user_voices;
   }else{
-    voices = app.bot_utils.get_uservoices_list();
+    voices = app.data_utils.get_uservoices_list();
   }
 
   let voice = { voice: app.voice_list[0].value, speed: 100, pitch: 100, intonation: 100, volume: 100, is_force_server: false };
@@ -30,12 +30,12 @@ module.exports = async (interaction, type) => {
   voices[member_id] = voice;
 
   if(!is_global_uservoice || type === 'is_force_server'){
-    app.bot_utils.write_serverinfo(guild_id, server_file, { user_voices: voices });
+    app.data_utils.write_serverinfo(guild_id, server_file, { user_voices: voices });
 
     const connection = app.connections_map.get(guild_id);
     if(connection) connection.user_voices = voices;
   }else{
-    app.bot_utils.write_uservoices_list(voices);
+    app.data_utils.write_uservoices_list(voices);
     app.setup_uservoice_list();
   }
